rabbit_feature_flags: Force feature flags using $RABBITMQ_FEATURE_FLAGS
If the environment variable is undefined, the default behavior is used:
all feature flags are enabled by default, except if the node is
clustered.

If the variable is defined to the empty string, no feature flag is
enabled.

If the variable is a list of feature flags (strings separated by comas),
only those feature flags are enabled if the node is unclustered.

The environment variable is undocumented on purpose (at least for now).